==> southeastern-europe/notebooks/01_elevation.ipynb <==
# 01: Pre-process elevation data

*Develop the elevation data which we will use in subsequent relative humidity and solar radiation calculations.*

import pygmt

Load in elevation dataset from pygmt.

ds = pygmt.datasets.load_earth_relief(
    resolution="15m", registration="pixel", region=[0, 360, -90, 90]
).to_dataset()

Because of the spatial resolution, many coastal grid cells will have erroneously negative temperatures, so clip negative elevations to zero. The vast majority of occurances of this effect are in coastal places, but it does also influence inland places that are below sea level.

ds = ds.where(ds > 0, 0)

ds["elevation"].plot()

Write out to bucket for use in subsequent steps.

output = "s3://carbonplan-climate-impacts/extreme-heat/v1.0/inputs/elevation.zarr"

ds.to_zarr(output, consolidated=True, mode="w")
